+++ src/main/java/controller/IndexController.java
@@ -3,35 +3,22 @@ package controller;
 import http.request.HttpRequest;
 import http.ContentType;
 import http.response.HttpResponse;
-import http.Uri;
 
 import java.io.IOException;
 import java.util.Collections;
-import java.util.List;
 
 import static utils.FileIoUtils.loadFile;
 
-public class IndexController implements Controller {
+public class IndexController extends AbstractController {
 
     private static final String INDEX_PATH = "/index.html";
 
-    private final List<String> paths;
-
-    public IndexController(){
+    public IndexController() {
         this.paths = Collections.singletonList("/");
     }
-    @Override
-    public void service(HttpRequest httpRequest, HttpResponse httpResponse) throws IOException {
-        doGet(httpResponse);
-    }
 
     @Override
-    public boolean isUri(HttpRequest httpRequest) {
-        Uri uri = httpRequest.getUri();
-        return paths.stream().anyMatch(uri::isEndsWith);
-    }
-
-    public void doGet(HttpResponse httpResponse) throws IOException {
+    public void doGet(HttpRequest httpRequest, HttpResponse httpResponse) throws IOException {
         String path = INDEX_PATH;
         ContentType contentType = ContentType.from(path);
         String filePath = contentType.getDirectory() + path;
