+++ src/main/java/controller/UserCreateController.java
@@ -9,32 +9,17 @@ import service.UserService;
 
 import java.io.IOException;
 import java.util.Collections;
-import java.util.List;
 import java.util.Map;
 
-public class UserCreateController implements Controller {
+public class UserCreateController extends AbstractController {
 
     private static final Logger logger = LoggerFactory.getLogger(UserCreateController.class);
-
-    private final List<String> paths;
 
     public UserCreateController() {
         this.paths = Collections.singletonList("/user/create");
     }
-    @Override
-    public void service(HttpRequest httpRequest, HttpResponse httpResponse) throws IOException {
-        String httpMethod = httpRequest.getMethod().toString();
-        if(httpMethod.equals("GET")) {
-            doGet(httpRequest, httpResponse);
-        }
-    }
 
     @Override
-    public boolean isUri(HttpRequest httpRequest) {
-        Uri uri = httpRequest.getUri();
-        return paths.stream().anyMatch(uri::isStartsWith);
-    }
-
     public void doGet(HttpRequest httpRequest, HttpResponse httpResponse) throws IOException {
         Uri uri = httpRequest.getUri();
         Map<String, String> queryParams = uri.getParameters();
@@ -42,5 +27,6 @@ public class UserCreateController implements Controller {
 
         httpResponse.sendRedirect("/index.html");
     }
+
 
 }
